Rename server.py to backend.py
